src/main/java/br/gov/servicos/editor/usuarios/TokenRecuperacaoSenhaService.java
package br.gov.servicos.editor.usuarios;

import br.gov.servicos.editor.usuarios.cadastro.TokenRecuperacaoSenhaRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Component;

import java.time.Clock;
import java.time.LocalDateTime;

import static java.lang.Long.valueOf;

@Component
public class TokenRecuperacaoSenhaService {

    public static final int TOKEN_LENGHT = 50;

    @Autowired
    private GeradorToken geradorToken;

    @Autowired
    private PasswordEncoder passwordEncoder;

    @Autowired
    private TokenRecuperacaoSenhaRepository repository;

    private Clock clock = Clock.systemUTC();

    public String gerarParaUsuario(String usuarioId) {
        String token = geradorToken.gerar();
        TokenRecuperacaoSenha tokenRecuperacaoSenha = new TokenRecuperacaoSenha()
                                                            .withUsuarioId(valueOf(usuarioId))
                                                            .withDataCriacao(LocalDateTime.now(clock))
                                                            .withToken(passwordEncoder.encode(token));
        repository.save(tokenRecuperacaoSenha);
        return token;
    }
}
